refactor: modularize server and websocket
