+++ src/httpServer.ts
@@ -1,0 +1,8 @@
+import http from 'http';
+import app from './app';
+
+const createHttpServer = (): http.Server => {
+  return http.createServer(app);
+};
+
+export default createHttpServer;
